extractBodyFrom (#32)
This is an optional method that I have found works great for extracting the body of a message. It does not change any of your existing code. You simply call $message->extractBodyFrom . This was written by buckfuddey.  See issue #30   It has worked reliably for me in production for 2 days now.   Admittedly not the style you have, but it works to patch up some of the holes that keep creating this error. I propose you add it for now until at which point later you can clean it up, if you wish.